requirement.txt
livekit-agents[silero,turn-detector,simplismart,openai]~=1.5

aiocsv
python-dotenv
gspread
google-auth

